+++ mpa/det/__init__.py
@@ -9,6 +9,9 @@ from . import inferrer
 from . import stage
 from . import trainer
 
+from . import incremental
+from . import semisl
+
 import mpa.modules.datasets.pipelines.torchvision2mmdet
 
 import mpa.modules.datasets.det_csv_dataset
