+++ mpa/det/exporter.py
@@ -12,13 +12,13 @@ from mmcv.runner import load_checkpoint
 
 from mpa.registry import STAGES
 from mpa.utils.logger import get_logger
-from .stage import DetectionStage
+from mpa.det.incremental import IncrDetectionStage
 
 logger = get_logger()
 
 
 @STAGES.register_module()
-class DetectionExporter(DetectionStage):
+class DetectionExporter(IncrDetectionStage):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
@@ -47,7 +47,7 @@ class DetectionExporter(DetectionStage):
                 model = model.cpu()
             precision = kwargs.pop('precision', 'FP32')
             logger.info(f'Model will be exported with precision {precision}')
-            
+
             export_model(model, cfg, output_path, target='openvino', precision=precision)
         except Exception as ex:
             # output_model.model_status = ModelStatus.FAILED
