+++ mpa/det/incremental/__init__.py
@@ -1,0 +1,7 @@
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from .stage import IncrDetectionStage
+
+__all__ = ["IncrDetectionStage"]
