+++ mpa/det/semisl/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (C) 2022 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+from . import trainer
+from . import inferrer
