+++ recipes/stages/_base_/data/coco_ubt.py
@@ -18,22 +18,9 @@ data = dict(
     workers_per_gpu=2,
     train=dict(
         type=__dataset_type,
-        super_type='PseudoBalancedDataset',
         ann_file=__data_root_path + 'annotations/instances_train2017.json',
         img_prefix=__data_root_path + 'train2017/',
         pipeline=__train_pipeline,
-        classes=[
-                'person', 'bicycle', 'car', 'motorcycle', 'airplane', 'bus', 'train', 'truckl',
-                'boat', 'traffic light', 'fire hydrant', 'stop sign', 'parking meter', 'bench', 'bird',
-                'cat', 'dog', 'horse', 'sheep', 'cow', 'elephant', 'bear', 'zebra', 'giraffe', 'backpack',
-                'umbrella', 'handbag', 'tie', 'suitcase', 'frisbee', 'skis', 'snowboard', 'sports ball',
-                'kite', 'baseball bat', 'baseball glove', 'skateboard', 'surfboard', 'tennis racket', 'bottle',
-                'wine glass', 'cup', 'fork', 'knife', 'spoon', 'bowl', 'banana', 'apple', 'sandwich', 'orange',
-                'broccoli', 'carrot', 'hot dog', 'pizza', 'donut', 'cake', 'chair', 'couch', 'potted plant', 'bed',
-                'dining table', 'toilet', 'tv', 'laptop', 'mouse', 'remote', 'keyboard', 'cell phone', 'microwave',
-                'oven', 'toaster', 'sink', 'refrigerator', 'book', 'clock', 'vase', 'scissors', 'teddy bear',
-                'hair drier', 'toothbrush'
-            ]
     ),
     unlabeled=dict(
         type=__dataset_type,
